app-back/requirements.txt
alembic==1.7.3
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cors==1.0.1
docopt==0.4.0
filelock==3.3.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
future==0.18.2
gevent==21.8.0
greenlet==1.1.2
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
mandrill==1.0.60
MarkupSafe==2.0.1
marshmallow==3.13.0
mysqlclient==2.0.3
pybin==0.3.4
PyJWT==1.4.2
PySocks==1.7.1
requests==2.26.0
requests-file==1.5.1
six==1.16.0
SQLAlchemy==1.4.25
tldextract==3.1.2
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
